Close OutputStream after writing log to file
I ran into some issues with the log file generated by the plugin on Windows.

When trying to clean the workspace before the build the generated console file from the previous build could not be deleted. This was the error:.
[Error 32] The process cannot access the file because it is being used by another process:

After restarting Jenkins and building again I was able to clean the workspace on the first build but all subsequent builds failed with the same error. 

I modified the source to manually close the output stream, installed it, and it appears to be working.